Lecture 6 HW fixes
• States translated to context to avoid re-renders
• Created menuItem component
• Added minor CSS adjustments
