+++ src/counter/component.jsx
@@ -1,12 +1,13 @@
 import { useState } from "react";
 import styles from './style.module.css'
 
+
+const itemsMin = 0;
+const itemsMax = 99;
+
 function Counter() {
     // Create a counter for added dishes
     const [itemsCounter, setItemsCounter] = useState(0);
-
-    const itemsMin = 0;
-    const itemsMax = 99;
 
     function addItem() {
         if (itemsCounter < itemsMax) {
