+++ src/menu/component.jsx
@@ -1,22 +1,16 @@
-import Counter from "../counter/component";
-import { useUser } from "../userContext";
+import MenuItem from "../menuItem/component";
 
 import styles from './style.module.css'
 
-
-
 function Menu({menu}) {
-    const userValue = useUser()
 
     return (  
     <div className={styles.menuWrapper}>
         <h3>Меню</h3>
         {menu?.length && (
-            <ul >
+            <ul className={styles.listWrapper}>
             {menu.map((element) => {
-                return <li className={styles.menuItem} key={element.id}>{element.name} 
-                {userValue.value !== null ? <Counter /> : ''}
-                </li>
+                return <MenuItem menuItem={element} />
             })}
         </ul>
         )}
