+++ src/menuItem/style.module.css
@@ -1,0 +1,7 @@
+
+.menuItem{
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+
+    justify-items: end;
+}
