+++ src/pageScroller/style.module.css
@@ -1,4 +1,4 @@
-#pageScroller{
+.pageScroller{
     position: fixed;
 
     height: 10px;
