+++ src/style.css
@@ -1,4 +1,4 @@
-*{
+body{
     margin: 0;
     padding: 0;
