+++ src/themeContext/index.jsx
@@ -1,0 +1,13 @@
+import { createContext, useContext, useState, useCallback } from "react"
+
+const ThemeContext = createContext()
+
+export const useTheme = () => useContext(ThemeContext)
+
+export const ThemeContextProvider = ({children})=>{
+
+    const [theme, setTheme] = useState('light')
+    const toggleTheme = useCallback(()=>{setTheme((current)=> current==='light'?'dark':'light')}, [])
+
+    return <ThemeContext.Provider value={{value: theme, toggleTheme}}>{children}</ThemeContext.Provider>
+};
